chapolin codigo

+++ structures/classFileStructures.h
@@ -12,7 +12,7 @@
 typedef unsigned char u1;
 typedef unsigned short u2;
 typedef unsigned int u4;
-typedef unsigned double u8;
+typedef unsigned long long u8;
 
 
 /*---------------------------------------------------------
